Add NotifierBase class
