+++ CoreLab/Core/Core.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Pipeline\Usual.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reactive\Notifier.cs" />
+    <Compile Include="Reactive\NotifierBase.cs" />
     <Compile Include="Reactive\ObservableEvent.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
